Fixed text in invite email
